pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "mlops_group54_project"
version = "0.0.1"
description = "mlops project 2026"
authors = [
  { name = "fiona_wennberg", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "hydra-core>=1.3.2",
    "torch==2.6.0",
    "torchvision>=0.21.0",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "wandb>=0.23.1",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]
